+++ src/utils.js
@@ -1,15 +1,14 @@
 const Debug = require('debug');
 const { Buffer } = require('buffer');
 const BN = require('bn.js');
-const Big = require('big.js');
 const JSZip = require('jszip');
 const NodeRSA = require('node-rsa');
 const aesjs = require('aes-js');
 const {
   getAddress,
   randomBytes,
-  formatEther,
-  parseEther,
+  formatUnits,
+  parseUnits,
 } = require('ethers').utils;
 const { BigNumber } = require('ethers');
 const multiaddr = require('multiaddr');
@@ -52,37 +51,56 @@ const stringify = (val) => {
 
 const formatRLC = (nRLC) => {
   try {
-    Big.NE = -10;
-    Big.PE = 10;
-    return new Big(stringify(nRLC)).times(new Big(10).pow(-9)).toString();
+    return formatUnits(stringify(nRLC), 9);
   } catch (error) {
     debug('formatRLC()', error);
     throw Error('Invalid nRLC');
   }
 };
 
-const parseRLC = (rlc) => {
+const parseRLC = (value, defaultUnit = 'RLC') => {
+  const [amount, inputUnit] = stringify(value).split(' ');
+  const unit = inputUnit !== undefined ? inputUnit : defaultUnit;
+  let pow;
+  if (unit === 'RLC') {
+    pow = 9;
+  } else if (unit === 'nRLC') {
+    pow = 0;
+  } else {
+    throw Error('Invalid token unit');
+  }
   try {
-    Big.NE = -10;
-    Big.PE = 18;
-    const rlcAmount = new Big(stringify(rlc));
-    return new BN(rlcAmount.times(new Big(10).pow(9)).toString());
+    return ethersBnToBn(parseUnits(amount, pow));
   } catch (error) {
     debug('parseRLC()', error);
-    throw Error('Invalid rlcString');
+    throw Error('Invalid token amount');
   }
 };
 
 const formatEth = (wei) => {
   try {
-    return formatEther(BigNumber.from(stringify(wei)));
+    return formatUnits(BigNumber.from(stringify(wei)));
   } catch (error) {
     debug('formatEth()', error);
     throw Error('Invalid wei');
   }
 };
 
-const parseEth = ether => ethersBnToBn(parseEther(stringify(ether)));
+const parseEth = (value, defaultUnit = 'ether') => {
+  const [amount, inputUnit] = stringify(value).split(' ');
+  const unit = inputUnit !== undefined ? inputUnit : defaultUnit;
+  if (
+    !['wei', 'kwei', 'mwei', 'gwei', 'szabo', 'finney', 'ether'].includes(unit)
+  ) {
+    throw Error('Invalid ether unit');
+  }
+  try {
+    return ethersBnToBn(parseUnits(amount, unit));
+  } catch (error) {
+    debug('formatEth()', error);
+    throw Error('Invalid ether amount');
+  }
+};
 
 const truncateBnWeiToBnNRlc = (bnWei) => {
   const weiString = bnWei.toString();